fix(fleet): move supervisor to own folder and split operation
